fix: 메인화면 API reload
